+++ plugins/source/aws/resources/services/cloudwatchlogs/metric_filters.go
@@ -1,11 +1,6 @@
 package cloudwatchlogs
 
 import (
-	"context"
-
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/aws/arn"
-	"github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs"
 	"github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -32,31 +27,4 @@ func MetricFilters() *schema.Table {
 			},
 		},
 	}
-}
-
-func fetchCloudwatchlogsMetricFilters(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	var config cloudwatchlogs.DescribeMetricFiltersInput
-	c := meta.(*client.Client)
-	svc := c.Services().Cloudwatchlogs
-	paginator := cloudwatchlogs.NewDescribeMetricFiltersPaginator(svc, &config)
-	for paginator.HasMorePages() {
-		page, err := paginator.NextPage(ctx)
-		if err != nil {
-			return err
-		}
-		res <- page.MetricFilters
-	}
-	return nil
-}
-
-func resolveMetricFilterArn(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
-	cl := meta.(*client.Client)
-	a := arn.ARN{
-		Partition: cl.Partition,
-		Service:   "logs",
-		Region:    cl.Region,
-		AccountID: cl.AccountID,
-		Resource:  "metric_filter/" + aws.ToString(resource.Item.(types.MetricFilter).FilterName),
-	}
-	return resource.Set(c.Name, a.String())
 }
